+++ packages/@expo/cli/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - fix launching apps on Android emulators where `hw.keyboard = no` ([#34677](https://github.com/expo/expo/pull/34677) by [@DimitarNestorov](https://github.com/DimitarNestorov))
 - only skip dependency validation for `EXPO_NO_DEPENDENCY_VALIDATION=1` ([#40043](https://github.com/expo/expo/pull/40043) by [@kitten](https://github.com/kitten))
+- Allow relative `export:embed` flags for Windows and monorepos ([#40563](https://github.com/expo/expo/pull/40563) by [@byCedric](https://github.com/byCedric))
 
 ### 💡 Others